refactor: cria custom hook
